Allow the controller to issue iframes on the browser.
The browser seems to spend about 25-30% of the time waiting for
the HTTP response for each iframe load (reusing the same port each
time), so that's about the odds we have of our payload getting
executed when we cross the SEQ boundry. If we lose, the browser
silently RSTs, and uses a new port.

I can't think of a good way to fix this, though I would very much
like something more elegant than "timeout and try again and again".

Note: this will also not work so great with a website that directly
sends a X-Frame-Options: DENY HTTP header, because browsers like to
kill the connection as soon as they get it. Note if the server redirects
to a page that does that (e.g. facebook -> https), you'll still be ok.