Make DoS use transitional windows
Instead of just jumping from 1KB table to 2KB, insert
a smooth transition table in between, where every other column
of the 1KB increases twice as fast. That way, it catches up
with the column to its right within one table, and the next table
can drop half the columns.